% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChunkData.R
\name{chunkData}
\alias{chunkData}
\title{Chunk the raw data}
\usage{
chunkData(connectionDetails, sourceDatabaseSchema, oracleTempSchema = NULL,
  nSubstring = 1)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{sourceDatabaseSchema}{The name of the database schema that contains the raw data
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'raw_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{nSubstring}{The number of digits to substring the person identifier by. Default is 1.}
}
\description{
Take the first n digits of the person identifier to chunk the raw data
}
