% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFinalVisitIDsTable.R
\name{createFinalVisitIDsTable}
\alias{createFinalVisitIDsTable}
\title{Run the query to create the assign_all_visit_ids table.}
\usage{
createFinalVisitIDsTable(connectionDetails, cdmDatabaseSchema,
  sourceDatabaseSchema, oracleTempSchema = NULL, iteraterNum)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{sourceDatabaseSchema}{The name of the database schema that contains the raw data
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'raw_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{iteraterNum}{The number of chunks of data the program should be looped over.}
}
\description{
This function will call and run the sql file to create the assign_all_visit_ids table which is the second step to creating the visit_occurrence table.
Requires that the function \code{chunkData} is run first.
}
