% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLookupTable.R
\name{createLookupTable}
\alias{createLookupTable}
\title{Run the query to create a lookup table.}
\usage{
createLookupTable(connectionDetails, cdmDatabaseSchema, sourceDatabaseSchema,
  sqlFile, oracleTempSchema = NULL)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{sourceDatabaseSchema}{The name of the database schema that contains the raw data
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'raw_instance.dbo'.}

\item{sqlFile}{The name of the sql file containing the sql to run. For example "LocationTable.sql"}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}
}
\description{
This function will call and run one sql file.
}
