% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVisitOccurrenceTable.R
\name{createVisitOccurrenceTable}
\alias{createVisitOccurrenceTable}
\title{Run the query to create the assign_all_visit_ids table.}
\usage{
createVisitOccurrenceTable(connectionDetails, cdmDatabaseSchema,
  sourceDatabaseSchema, oracleTempSchema = NULL, runAllParts = TRUE,
  iteraterNum)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{sourceDatabaseSchema}{The name of the database schema that contains the raw data
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'raw_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}

\item{runAllParts}{Set this parameter as TRUE to run all parts of the VISIT_OCCURRENCE table. If set
to FALSE it will run only the final portion. The first three pieces
are available as separate functions called \code{createAllVisitTable}
and \code{createFinalVisitIDsTable}.}
}
\description{
This function will call and run the sql file to create the visit_occurrence table which is the second step to creating the visit_occurrence table.
Requires that the function \code{chunkData} is run first.
}
