% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVocabTables.R
\name{createVocabTables}
\alias{createVocabTables}
\title{Run the vocab queries to create the vocab mapping tables.}
\usage{
createVocabTables(connectionDetails, cdmDatabaseSchema, vocabDatabaseSchema,
  oracleTempSchema = NULL)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabDatabaseSchema}{The name of the database schema that contains the OMOP CDM Vocabulary
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'vocab_instance.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}
}
\description{
Run the vocab queries to create the vocab mapping tables.
}
