% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIteraterNum.R
\name{getIteraterNum}
\alias{getIteraterNum}
\title{Get the number of chunks the program will need to loop over}
\usage{
getIteraterNum(list = listHVIDChunks)
}
\description{
Get the number of chunks the program will need to loop over
}
